test(board): keep the /board route test from dragging every page into coverage
The frontend CI coverage gate regressed (82% -> 80.5% lines) because
BoardRoute.test.tsx imported AppRouter, whose static page imports pulled
untested pages (AccountHome, PresetsPage, AdminPage, ...) into the v8
coverage denominator for the first time.

Extract the board route into its own modules so it can be imported —
and tested — in isolation:

* src/pages/BoardPage.tsx — the Board component (credential-mode
  resolution, per-board PWA manifest pointing, owner-bookmark upgrade).
* src/auth/RequireAuth.tsx — RequireAuth + the shared RouteLoading
  placeholder, out of AppRouter for the same reason.

AppRouter keeps PublicOnly and the route table. The test now renders
BoardPage inside a MemoryRouter with a /login probe, which also adds
coverage for the login bounce preserving the destination.
