+++ frontend/src/test/BoardRoute.test.tsx
@@ -1,11 +1,14 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, waitFor } from '@testing-library/react';
-import AppRouter from '../AppRouter';
+import { MemoryRouter, Route, Routes, useLocation } from 'react-router-dom';
+import { AuthProvider } from '../auth/AuthContext';
+import BoardPage from '../pages/BoardPage';
 import * as api from '../api/client';
 
 // The board itself is exercised by App.test.tsx — here only the /board route
-// wiring matters: which credential mode Board hands to App and where it points
-// the PWA manifest. A stub that echoes its props keeps the test hermetic.
+// wiring matters: which credential mode BoardPage hands to App and where it
+// points the PWA manifest. A stub that echoes its props keeps the test
+// hermetic (and keeps the heavyweight App out of the module graph).
 vi.mock('../App', () => ({
   default: (props: { controlToken?: string; publicUser?: string }) => (
     <div
@@ -48,6 +51,28 @@ function loggedInAs(username: string) {
   };
 }
 
+/** Stub /login that records the location state RequireAuth forwarded. */
+function LoginProbe() {
+  const location = useLocation();
+  const from = (location.state as { from?: { pathname?: string; search?: string } } | null)?.from;
+  return (
+    <div data-testid="login-page" data-from={`${from?.pathname ?? ''}${from?.search ?? ''}`} />
+  );
+}
+
+function renderBoard(path: string) {
+  return render(
+    <MemoryRouter initialEntries={[path]}>
+      <AuthProvider>
+        <Routes>
+          <Route path="/board" element={<BoardPage />} />
+          <Route path="/login" element={<LoginProbe />} />
+        </Routes>
+      </AuthProvider>
+    </MemoryRouter>,
+  );
+}
+
 function manifestLink(): HTMLLinkElement {
   const link = document.querySelector<HTMLLinkElement>('link[rel="manifest"]');
   if (!link) throw new Error('manifest link missing');
@@ -62,8 +87,7 @@ describe('/board route', () => {
 
   it('keeps bookmark mode for an anonymous ?u= visitor and points the manifest at it', async () => {
     vi.mocked(api.getAuthContext).mockResolvedValue(anonymous);
-    window.history.pushState({}, '', '/board?u=alex&oid=liga');
-    render(<AppRouter />);
+    renderBoard('/board?u=alex&oid=liga');
 
     const app = await screen.findByTestId('board-app');
     expect(app.dataset.publicUser).toBe('alex');
@@ -72,8 +96,7 @@ describe('/board route', () => {
 
   it('upgrades an own ?u= bookmark to owner mode when the owner is signed in', async () => {
     vi.mocked(api.getAuthContext).mockResolvedValue(loggedInAs('alex'));
-    window.history.pushState({}, '', '/board?u=Alex&oid=liga');
-    render(<AppRouter />);
+    renderBoard('/board?u=Alex&oid=liga');
 
     const app = await screen.findByTestId('board-app');
     // Usernames are stored lowercased; the URL casing must not defeat the match.
@@ -82,8 +105,7 @@ describe('/board route', () => {
 
   it('keeps bookmark mode when a different account is signed in', async () => {
     vi.mocked(api.getAuthContext).mockResolvedValue(loggedInAs('sam'));
-    window.history.pushState({}, '', '/board?u=alex&oid=liga');
-    render(<AppRouter />);
+    renderBoard('/board?u=alex&oid=liga');
 
     const app = await screen.findByTestId('board-app');
     expect(app.dataset.publicUser).toBe('alex');
@@ -91,8 +113,7 @@ describe('/board route', () => {
 
   it('points the manifest at the owner board for a plain ?oid= visit', async () => {
     vi.mocked(api.getAuthContext).mockResolvedValue(loggedInAs('alex'));
-    window.history.pushState({}, '', '/board?oid=liga');
-    render(<AppRouter />);
+    renderBoard('/board?oid=liga');
 
     await screen.findByTestId('board-app');
     await waitFor(() =>
@@ -100,10 +121,19 @@ describe('/board route', () => {
     );
   });
 
+  it('bounces an anonymous ?oid= visit to /login, preserving the destination', async () => {
+    vi.mocked(api.getAuthContext).mockResolvedValue(anonymous);
+    renderBoard('/board?oid=liga');
+
+    const login = await screen.findByTestId('login-page');
+    // RequireAuth stashes the origin so LoginPage can navigate back — this is
+    // what makes an installed /board?oid= PWA usable without a live session.
+    expect(login.dataset.from).toBe('/board?oid=liga');
+  });
+
   it('leaves the manifest alone for a revocable ?c= link', async () => {
     vi.mocked(api.getAuthContext).mockResolvedValue(anonymous);
-    window.history.pushState({}, '', '/board?c=tok123&oid=liga');
-    render(<AppRouter />);
+    renderBoard('/board?c=tok123&oid=liga');
 
     const app = await screen.findByTestId('board-app');
     expect(app.dataset.controlToken).toBe('tok123');
